+++ src/lib/formatters.ts
@@ -2,12 +2,10 @@ import {
   clone,
   find,
   first,
-  flatMap,
   groupBy,
   last,
   omit,
   sortBy,
-  uniqBy,
   zipObject,
 } from 'lodash-es';
 import moment from 'moment';
@@ -495,6 +493,28 @@ export function formatTimetableLabel(timetable) {
 
   return timetableLabel;
 }
+
+export const formatRouteName = (route: Record<string, string>) => {
+  if (route.route_long_name === null || route.route_long_name === '') {
+    return `Route ${route.route_short_name}`;
+  }
+
+  return route.route_long_name ?? 'Unknown';
+};
+
+export const formatRouteNames = (routes: Record<string, string>[]) => {
+  return new Intl.ListFormat('en-US', {
+    style: 'long',
+    type: 'conjunction',
+  }).format(routes.map((route) => formatRouteName(route)));
+};
+
+export const formatAgencyNames = (agencies: { agency_name: string }[]) => {
+  return new Intl.ListFormat('en-US', {
+    style: 'long',
+    type: 'conjunction',
+  }).format(agencies.map((agency) => agency.agency_name));
+};
 
 /*
  * Merge timetables with same `timetable_id`.